+++ apps/src/lab2/levelEditors/predictSettings/EditPredictSettings.tsx
@@ -2,10 +2,10 @@ import React, {useState} from 'react';
 import {LevelPredictSettings, PredictQuestionType} from '../types';
 import {SimpleDropdown} from '@cdo/apps/componentLibrary/dropdown';
 import Checkbox from '@cdo/apps/componentLibrary/checkbox';
-import TextField from '@cdo/apps/componentLibrary/textField/TextField';
 import moduleStyles from './edit-predict-settings.module.scss';
 import {BodyThreeText} from '@cdo/apps/componentLibrary/typography';
-import {Button} from '@cdo/apps/componentLibrary/button';
+import FreeResponseFields from './FreeResponseFields';
+import MultipleChoiceFields from './MultipleChoiceFields';
 
 interface EditPredictSettingsProps {
   initialSettings: LevelPredictSettings;
@@ -36,81 +36,6 @@ const EditPredictSettings: React.FunctionComponent<
     });
   };
 
-  const handleToggleMultipleChoiceAnswer = (
-    e: React.ChangeEvent<HTMLInputElement>
-  ) => {
-    const newCorrectAnswers = predictSettings.multiple_choice_answers
-      ? [...predictSettings.multiple_choice_answers]
-      : [];
-    if (e.target.checked && !newCorrectAnswers.includes(e.target.value)) {
-      newCorrectAnswers.push(e.target.value);
-    } else if (
-      !e.target.checked &&
-      newCorrectAnswers.includes(e.target.value)
-    ) {
-      newCorrectAnswers.splice(newCorrectAnswers.indexOf(e.target.value), 1);
-    }
-    console.log({newCorrectAnswers});
-    setPredictSettings({
-      ...predictSettings,
-      multiple_choice_answers: newCorrectAnswers,
-    });
-  };
-
-  const renderFreeResponseOptions = () => {
-    return (
-      <div>
-        <TextField
-          label="Placeholder text to show in the answer text area"
-          value={predictSettings.placeholder_text}
-          onChange={e =>
-            setPredictSettings({
-              ...predictSettings,
-              placeholder_text: e.target.value,
-            })
-          }
-          name="placeholder_text"
-          className={moduleStyles.fieldArea}
-        />
-        <label className={moduleStyles.fieldArea}>
-          <div className={moduleStyles.label}>
-            Optional solution markdown to display to teachers
-          </div>
-          <textarea
-            value={predictSettings.solution}
-            onChange={e =>
-              setPredictSettings({
-                ...predictSettings,
-                solution: e.target.value,
-              })
-            }
-            name="solution"
-            rows={3}
-            className={'input-block-level'}
-          />
-        </label>
-        <label className={moduleStyles.fieldArea}>
-          <div className={moduleStyles.label}>
-            Height of the answer textarea in px
-          </div>
-          <input
-            type="number"
-            value={predictSettings.free_response_height}
-            onChange={e =>
-              setPredictSettings({
-                ...predictSettings,
-                free_response_height: parseInt(e.target.value),
-              })
-            }
-            name="free_response_height"
-            min={0}
-            step={50}
-          />
-        </label>
-      </div>
-    );
-  };
-
   const renderMultipleChoiceOptions = () => {
     if (!predictSettings.multiple_choice_options) {
       setPredictSettings({
@@ -119,76 +44,10 @@ const EditPredictSettings: React.FunctionComponent<
       });
     }
     return (
-      <div>
-        <label className={moduleStyles.fieldArea}>
-          <div className={moduleStyles.label}>Multiple Choice Options</div>
-          {predictSettings.multiple_choice_options &&
-            predictSettings.multiple_choice_options.map((option, index) => (
-              <div key={index} className={moduleStyles.multipleChoiceOption}>
-                <input
-                  type="text"
-                  value={option}
-                  onChange={e => {
-                    const newOptions = [
-                      ...predictSettings.multiple_choice_options!,
-                    ];
-                    newOptions[index] = e.target.value;
-                    setPredictSettings({
-                      ...predictSettings,
-                      multiple_choice_options: newOptions,
-                    });
-                  }}
-                  name={`multiple_choice_option_${index}`}
-                />
-                <Checkbox
-                  label="Mark as correct answer"
-                  checked={
-                    predictSettings.multiple_choice_answers?.includes(option) ||
-                    false
-                  }
-                  onChange={handleToggleMultipleChoiceAnswer}
-                  name={`mark_correct_answer_${index}`}
-                  value={option}
-                />
-                {index > 0 && (
-                  <Button
-                    onClick={() => {
-                      const newOptions = [
-                        ...predictSettings.multiple_choice_options!,
-                      ];
-                      newOptions.splice(index, 1);
-                      setPredictSettings({
-                        ...predictSettings,
-                        multiple_choice_options: newOptions,
-                      });
-                    }}
-                    ariaLabel={'Delete option'}
-                    color={'black'}
-                    size={'xs'}
-                    isIconOnly={true}
-                    icon={{iconName: 'trash'}}
-                  />
-                )}
-              </div>
-            ))}
-          <Button
-            onClick={() =>
-              setPredictSettings({
-                ...predictSettings,
-                multiple_choice_options: [
-                  ...predictSettings.multiple_choice_options!,
-                  '',
-                ],
-              })
-            }
-            isIconOnly={true}
-            color={'black'}
-            size={'s'}
-            icon={{iconName: 'plus'}}
-            ariaLabel={'Add Option'}
-          />
-        </label>
-      </div>
+      <MultipleChoiceFields
+        predictSettings={predictSettings}
+        setPredictSettings={setPredictSettings}
+      />
     );
   };
 
@@ -231,9 +90,15 @@ const EditPredictSettings: React.FunctionComponent<
             size="s"
             className={moduleStyles.fieldArea}
           />
-          {predictSettings.question_type === PredictQuestionType.FreeResponse
-            ? renderFreeResponseOptions()
-            : renderMultipleChoiceOptions()}
+          {predictSettings.question_type ===
+          PredictQuestionType.FreeResponse ? (
+            <FreeResponseFields
+              predictSettings={predictSettings}
+              setPredictSettings={setPredictSettings}
+            />
+          ) : (
+            renderMultipleChoiceOptions()
+          )}
           <Checkbox
             label="Allow multiple tries"
             checked={predictSettings.allow_multiple_attempts || false}
